feat: connect enrollment dashboard to backend and fix style issues
