+++ composer.json
@@ -23,7 +23,7 @@
         "larastan/larastan": "^3.0",
         "laravel/dusk": "^8.3",
         "laravel/pail": "^1.2.2",
-        "laravel/pint": "^1.18",
+        "laravel/pint": "^1.25",
         "laravel/sail": "^1.41",
         "mockery/mockery": "^1.6",
         "nunomaduro/collision": "^8.6",
